services/django/lithium/timetabling_system/api/serializers.py
from rest_framework import serializers
from timetabling_system.models import Exam

class ExamSerializer(serializers.ModelSerializer):
    class Meta:
        model = Exam
        fields = "__all__"
