services/django/lithium/pyproject.toml
[project]
name = "lithium"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "django ~=5.1",
  "django-allauth[socialaccount] ~=65.2",
  "django-debug-toolbar ~=4.4",
  "crispy-bootstrap5 ~=2024.10",
  "gunicorn ~=23.0",
  "psycopg[binary] ~=3.2",
  "whitenoise ~=6.7",
  "djangorestframework ~=3.15"
]

[dependency-groups]
dev = [
  "typing_extensions ~=4.12",
]
